the use of org.sonatype.oss:oss-parent:9 is discouraged/depreciated per http://central.sonatype.org/pages/apache-maven.html
"Deprecated oss-parent

In the past all the plugin configuration and other setup was managed by a Maven parent POM with the latest coordinates of org.sonatype.oss:oss-parent:9. This project leaked SCM, URL and other details and its usage is discouraged. Maintenance of the project has stopped and it no longer works with latest tooling such as Maven versions or Java versions. ..."